migrated query processing logic
